refactor(x): collapse the 12 view booleans in App.tsx onto ViewState
App.tsx modelled "which view is showing" twice: correctly as the ViewState
union (used only for back/forward history and deep links), and badly as 12
mutually-exclusive useState booleans hand-synced across ~700 setter calls,
plus 13 sites that enumerated the whole set in a &&/|| chain. The render is
a ternary chain, so two-true silently rendered the first match and all-false
rendered a blank pane. Neither errored.

There is now one useState<ViewState>, and every is*Open flag plus
selectedPath and selectedBackgroundTask are derived from it, so both states
are unrepresentable. Readers are untouched; the setters collapse to 24
setView calls, and the 13 enumerating guards become three named predicates.
isTabbedView and isTabbedViewExceptCode differ only by the Code view, which
several guards predate - folding them together would change behaviour.

isBrowserOpen stays a separate boolean: it overlays whatever view is
showing rather than being one. workspaceInitialPath, knowledgeViewFolderPath
and knowledgeViewMode also stay separate, because they deliberately outlive
the view - switching away from the workspace tab and back must land on the
folder being browsed, not the root.

New viewForTabPath() replaces the 12-branch tab-path-to-view chain that
switchFileTab and closeFileTab each hand-rolled. That fixes a latent bug:
closeFileTab had no isCodeTabPath branch, so closing a tab beside the Code
tab set selectedPath to the sentinel '__rowboat_code__'. It rendered
correctly only because isCodeOpen won the ternary chain, while history
recorded a bogus file path.

Also lifts the pure helpers App.tsx had accumulated into lib/ (knowledge-tree,
view-state, copilot-prompts, and appended to wiki-links and frontmatter).
7,704 -> 6,729 lines. Typecheck clean; renderer suite 134 passing, from 119.
